・serializersのインストール ・javascriptを使ってコメント情報を表示 ・コメント追加ボタンとテキストフォームの作成
