+++ app/assets/stylesheets/form.scss
@@ -18,4 +18,7 @@ input,select {
     min-height: 440px;
     resize: vertical;
     border: 1px solid #EBEBEB;
+    &.sm {
+    min-height: 160px;
+    }
   }
