+++ app/controllers/articles_controller.rb
@@ -6,7 +6,6 @@ class ArticlesController < ApplicationController
     end
 
     def show
-        @comments = @article.comments
     end
 
     def new
