+++ app/controllers/comments_controller.rb
@@ -4,6 +4,12 @@ class CommentsController < ApplicationController
         @comment = article.comments.build
     end
 
+    def index
+      article = Article.find(params[:article_id])
+      comments = article.comments
+      render json: comments
+    end
+
     def create
       article = Article.find(params[:article_id])
       @comment = article.comments.build(comment_params)
