+++ app/serializers/comment_serializer.rb
@@ -1,0 +1,3 @@
+class CommentSerializer < ActiveModel::Serializer
+  attributes :id, :content
+end
