+++ app/views/articles/show.html.haml
@@ -26,11 +26,15 @@
 
 .article
   %h2 コメント一覧
-  - @comments.each do |comment|
-    .article_comment
-      %p= comment.content
+  .comments-container
 
 .container
-  = link_to new_article_comment_path(@article) do
+  .btn-secondary.show-comment-form
+    コメントを追加
+  .comment-text-area.hidden
+    = text_area_tag :comment_content, '', class: 'sm'
     .btn-secondary
       コメントを追加
+
+
+= javascript_pack_tag 'article'
