+++ config/routes.rb
@@ -5,7 +5,7 @@ Rails.application.routes.draw do
   resource :timeline, only: [:show]
 
   resources :articles do
-      resources :comments, only: [:new, :create]
+      resources :comments, only: [:index, :new, :create]
 
       resource :like, only: [:show, :create, :destroy]
     end
